Move back to uint64

+++ experiments/tripreport_reduced.py
@@ -5,32 +5,32 @@ def gen_schema(file, value_max=np.iinfo(np.uint32).max, value_min=0):
     import pyarrow as pa
 
     schema_fields = [pa.field("timestamp", pa.utf8(), False),
-                     pa.field("timezone", pa.uint32(), False).with_metadata(
+                     pa.field("timezone", pa.uint64(), False).with_metadata(
                          {"illex_MIN": str(value_min), "illex_MAX": str(value_max)}),
-                     pa.field("odometer", pa.uint32(), False).with_metadata(
+                     pa.field("odometer", pa.uint64(), False).with_metadata(
                          {"illex_MIN": str(value_min), "illex_MAX": str(value_max)}),
                      pa.field("hypermiling", pa.bool_(), False),
-                     pa.field("avgspeed", pa.uint32(), False).with_metadata(
+                     pa.field("avgspeed", pa.uint64(), False).with_metadata(
                          {"illex_MIN": str(value_min), "illex_MAX": str(value_max)}),
                      pa.field("sec_in_band",
-                              pa.list_(pa.field("item", pa.uint32(), False)
+                              pa.list_(pa.field("item", pa.uint64(), False)
                                        .with_metadata({"illex_MIN": str(value_min), "illex_MAX": str(value_max)}),
                                        12), False),
                      pa.field("braking", pa.list_(
-                         pa.field("item", pa.uint32(), False).with_metadata(
+                         pa.field("item", pa.uint64(), False).with_metadata(
                              {"illex_MIN": str(value_min), "illex_MAX": str(value_max)}),
                          6), False),
                      pa.field("accel", pa.list_(
-                         pa.field("item", pa.uint32(), False).with_metadata(
+                         pa.field("item", pa.uint64(), False).with_metadata(
                              {"illex_MIN": str(value_min), "illex_MAX": str(value_max)}),
                          6), False),
                      pa.field("orientation", pa.bool_(), False),
                      pa.field("small_speed_var", pa.list_(
-                         pa.field("item", pa.uint32(), False).with_metadata(
+                         pa.field("item", pa.uint64(), False).with_metadata(
                              {"illex_MIN": str(value_min), "illex_MAX": str(value_max)}),
                          13), False),
                      pa.field("large_speed_var", pa.list_(
-                         pa.field("item", pa.uint32(), False).with_metadata(
+                         pa.field("item", pa.uint64(), False).with_metadata(
                              {"illex_MIN": str(value_min), "illex_MAX": str(value_max)}),
                          13), False)
                      ]
